backend/requirements.txt
ansicon==1.89.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
blessed==1.20.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
cloudpickle==3.1.0
confluent-kafka==2.6.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-extensions==3.2.3
django-picklefield==3.2
django-prometheus==2.3.1
django-q==1.3.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
fsspec==2024.10.0
gevent==24.11.1
greenlet==3.1.1
gym==0.26.2
gym-notices==0.0.8
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
jinxed==1.3.0
kafka-python==2.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.3
pandas==2.2.3
pathlib==1.0.1
pillow==11.0.0
psutil==6.1.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
websockets==14.1
zope.event==5.0
zope.interface==7.2
